# web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "V1.5 web layer \u2014 Pydantic-driven contracts over the V1 + V1.5a engines. Localhost-only (no auth, V1.5).",
    "title": "SecBrain V1.5b",
    "version": "1.5.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/settings/feedback-loop": {
      "get": {
        "description": "Return the active ContextBlock for a (corpus, template) pair.\n\nUsed by `/settings/feedback-loop` UI to render the live context block +\nblocklist that Pass 4 will see on its next invocation.",
        "operationId": "view_feedback_loop_api_v1_settings_feedback_loop_get",
        "parameters": [
          {
            "in": "query",
            "name": "corpus_id",
            "required": true,
            "schema": {
              "title": "Corpus Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "prompt_template_id",
            "required": true,
            "schema": {
              "title": "Prompt Template Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response View Feedback Loop Api V1 Settings Feedback Loop Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "View Feedback Loop",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/feedback-loop/decisions": {
      "post": {
        "description": "Append a manual feedback decision (operator override).",
        "operationId": "append_decision_api_v1_settings_feedback_loop_decisions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Append Decision Api V1 Settings Feedback Loop Decisions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Append Decision",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "operationId": "list_sources_api_v1_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_disconnected",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Disconnected",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Sources Api V1 Sources Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sources",
        "tags": [
          "sources"
        ]
      },
      "post": {
        "description": "POST /api/v1/sources \u2014 connect_data_source.",
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Payload",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Source Api V1 Sources Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Source",
        "tags": [
          "sources"
        ]
      }
    },
...
  }
}